Move the client's notification routing into its own mixin
The four private methods that hook a transport's notifications up and
act on them -- dropping caches, running the host's listeners -- are one
unit, and moving them keeps Client under its length cop the same way the
listing and task mixins already do. Nothing about them changes.
